// src/interfaces/Leader.d.ts
// Generated by https://quicktype.io
//
// To change quicktype's target language, run command:
//
//   "Set quicktype target language"

export interface ILeader {
    Faction:             Faction;
    StartingArmyOptions: DisplayInfo[];
    Id:                  number;
    Name:                string;
    DisplayInfo:         DisplayInfo;
    Image:               TmpImage;
    HW2PromotionOffer:   HW2PromotionOffer;
}

export interface HW2StartingArmy {
    DisplayInfo: DisplayInfo;
    Cards:       Card[];
}

export interface DisplayInfoView {
    Status:                      Status;
    Common:                      Common;
    Identity:                    string;
    BatchLocalization:           BatchLocalization;
    Localization:                Localization;
    Title:                       string;
    HW2LeaderDisplayInfo?:       HW2LeaderDisplayInfo;
    HW2StartingArmy?:            HW2StartingArmy;
    Autoroute?:                  string;
    HW2StartingArmyDisplayInfo?: HW2StartingArmyDisplayInfo;
}

export interface DisplayInfo {
    Id:   number;
    Type: DisplayInfoType;
    View: DisplayInfoView;
}

export interface Card {
    Id:         number;
    Type:       string;
    Identity:   string;
    Autoroute?: string;
}


export interface BatchLocalization {
    IsLocked: boolean;
}

export interface Common {
    Owner:        any;
    CreatedUtc:   string;
    ModifiedUtc:  string;
    PublishedUtc: string;
    Container:    number;
}

export interface HW2LeaderDisplayInfo {
    Name:        string;
    Description: string;
}

export interface HW2StartingArmyDisplayInfo {
    Name: string;
}

export interface Localization {
    Culture:             Culture;
    MasterContentItemId: string;
}


export type DisplayInfoType = "HW2LeaderDisplayInfo"|"HW2StartingArmy"|"HW2StartingArmyDisplayInfo"


export type Faction = "Banished" | "UNSC"

export interface HW2PromotionOffer {
    Id?:        number;
    Type?:      CardType;
    Identity?:  string;
    Autoroute?: string;
}

export interface TmpImage {
    Id:   number;
    Type: ImageType;
    View: ImageView;
}

export interface ImageView {
    Status:   string;
    Common:   Common;
    Media:    Media;
    Image:    ViewImage;
    Title:    string;
    Identity: string;
}

export interface ViewImage {
    Width:  number;
    Height: number;
}

export interface Media {
    MediaUrl:      string;
    MediaHeadUrl: string;
    MediaHeadUrlMain?: string;
    MimeType:      string;
    Caption:       string;
    AlternateText: string;
    FolderPath:    FolderPath;
    FileName:      string;
}